# Add delta page cache