--- a/pageserver/src/big_cache.rs
+++ b/pageserver/src/big_cache.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, io::Write, path::Path};
+use std::{collections::HashMap, io::{Seek, SeekFrom, Write}, ops::Range, os::unix::prelude::FileExt, path::Path};
 
 use bytes::Bytes;
 use anyhow::Result;
@@ -12,25 +12,41 @@ use crate::{layered_repository::ephemeral_file::EphemeralFile, repository::Key};
 pub static BIG_CACHE: Lazy<BigCache> = Lazy::new(|| {
     BigCache {
         index: Mutex::new(HashMap::new()),
-        file: Mutex::new(File::create(Path::new("big_cache.tmp")).unwrap()),
+        file: Mutex::new(File::options()
+            .read(true)
+            .write(true)
+            .create(true)
+            .open(Path::new("big_cache.tmp")).unwrap()),
     }
 });
 
+// TODO rename this
 pub struct BigCache {
-    // TODO point to file offset instead
-    index: Mutex<HashMap<Key, (Lsn, Bytes)>>,
-    file: Mutex<File>,  // TODO use ephemeral file instead
+    // TODO more granular locks for parallel access
+    index: Mutex<HashMap<Key, (Lsn, Range<u64>)>>,
+    file: Mutex<File>,  // TODO use EphemeralFile
 }
 
 impl BigCache {
     pub fn lookup(&self, key: &Key) -> Result<Option<(Lsn, Bytes)>> {
-        // TODO read this from the file instead
-        Ok(self.index.lock().unwrap().get(key).map(|img| img.clone()))
+        match self.index.lock().unwrap().get(key).map(|img| img.clone()) {
+            Some((lsn, range)) => {
+                let mut buf = vec![0u8; (range.end - range.start) as usize];
+                self.file.lock().unwrap().read_exact_at(&mut buf, range.start)?;
+                Ok(Some((lsn, Bytes::copy_from_slice(&buf))))
+            }
+            None => Ok(None),
+        }
     }
 
+    // TODO add 10MB buffer
     pub fn memorize(&self, key: Key, img: (Lsn, Bytes)) -> Result<()> {
-        self.index.lock().unwrap().insert(key, img.clone());
-        self.file.lock().unwrap().write(&img.1)?;
+        let mut file = self.file.lock().unwrap();
+        let begin = file.seek(SeekFrom::Current(0))?;
+        file.write(&img.1)?;
+        let end = file.seek(SeekFrom::Current(0))?;
+
+        self.index.lock().unwrap().insert(key, (img.0, begin..end));
         Ok(())
     }
 }