As an experiment only, I'm caching all modified pages during checkpoint into an ephemeral file cache.

Results:
```
old:
test_pgbench[zenith-45-10].simple-update.latency_average: 0.669 ms
test_pgbench[zenith-45-10].simple-update.latency_stddev: 0.216 ms
test_pgbench[zenith-45-10].simple-update.tps: 5,958.3685
test_pgbench[zenith-45-10].select-only.latency_average: 0.365 ms
test_pgbench[zenith-45-10].select-only.latency_stddev: 0.621 ms
test_pgbench[zenith-45-10].select-only.tps: 10,958.0510

new:
test_pgbench[zenith-45-10].simple-update.latency_average: 0.630 ms
test_pgbench[zenith-45-10].simple-update.latency_stddev: 0.218 ms
test_pgbench[zenith-45-10].simple-update.tps: 6,327.8910
test_pgbench[zenith-45-10].select-only.latency_average: 0.212 ms
test_pgbench[zenith-45-10].select-only.latency_stddev: 0.160 ms
test_pgbench[zenith-45-10].select-only.tps: 18,705.8595
```
Summary: `simple-update` is not worse (surprising!), `select-only` is much better. I'll measure how much ephemeral file bloat this cache creates. I'm sure the [hot_page test](https://github.com/neondatabase/neon/pull/1479) is also much better but I haven't tested yet.

The `get_reconstruct_data` function still shows up at 11% on the flame graph. Not sure why. All pages should be either inmem or in the big cache.

The purpose of this draft PR is to share results, and maybe later arrive at something mergeable. If you have experiments you'd like me to run pls share.