@hlinnaka 2022-04-26
Hmm, we already have the materialized page cache. It's built lazily, but should have the same effect in the long run I think. Do you get similar results if you just bump up page_cache_size large enough that everything fits in cache? (The cache has to 2x - 3x larger than the dataset, because the replacement policy isn't great, see discussion at https://github.com/neondatabase/neon/pull/1530).