@bojanserafimov 2022-04-26
But yes, most of the gain is coming from extending the materialized page cache into disk, and not from eager materialization during checkpointing. Unless I find a way to remove that 10% of time spent in getting reconstruct data.