@bojanserafimov 2022-04-27
Eager materialization shows problems on larger workloads. Init is way too slow. Closing this, but might follow up with a PR extending the existing materialized page cache to a file.